\docType{package}
\name{RBSperformance-package}
\alias{RBSperformance}
\alias{RBSperformance-package}
\title{An R package to aid in RBS estimations}
\description{
  An R package to aid in RBS estimations
}
\author{
  Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}
}
\references{
  Gregoire, T G, H T Valentine, and G M Furnival. "Sampling
  Methods to Estimate Foliage and Other Characteristics of
  Individual Trees" Ecology 76 (1995): 1181-1194.

  Bascietto, Marco, Bruno De Cinti, Giorgio Matteucci, and
  Alessandro Cescatti. "Biometric Assessment of Aboveground
  Carbon Pools and Fluxes in Three European Forests by
  Randomized Branch Sampling" Forest Ecology and Management
  267 (March 2012): 172-181.

  Source code is hosted at GitHub
  (\url{http://mbask.github.com/RBSperformance/})
}
\keyword{package}

